% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{as.list.bdeconv_result}
\alias{as.list.bdeconv_result}
\title{Density deconvolution using bayesian semiparametric methods}
\usage{
\method{as.list}{bdeconv_result}(x, ...)
}
\arguments{
\item{x}{A \code{bdeconv_result} object.}

\item{...}{Unused.}
}
\description{
Density deconvolution using bayesian semiparametric methods
}
\keyword{internal}
