% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.bdeconv_result}
\alias{autoplot.bdeconv_result}
\title{Visualization of \code{bdeconv_result} object using \code{ggplot2}}
\usage{
autoplot.bdeconv_result(object)
}
\arguments{
\item{object}{A \code{bdeconv_result} object.}
}
\value{
A collection of plots illustrating various densities and related functions made using \code{ggplot2}.
}
\description{
Visualization of \code{bdeconv_result} object using \code{ggplot2}
}
