% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesdecon-package.R
\docType{package}
\name{BayesDecon-package}
\alias{BayesDecon}
\alias{BayesDecon-package}
\title{Density Deconvolution Using Bayesian Semiparametric Methods}
\description{
Placeholder...
}
\author{
Blake Moya \href{mailto:blakemoya@utexas.edu}{blakemoya@utexas.edu}
}
\keyword{internal}
