% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesdecon_data_simulated.R
\docType{data}
\name{BayesDecon_data_simulated}
\alias{BayesDecon_data_simulated}
\title{Simulated dataset for BayesDecon}
\format{
A data frame with 3000 rows and 4 variables.
}
\source{
Generated for package examples.
}
\usage{
BayesDecon_data_simulated
}
\description{
A small example dataset used to demonstrate bayesdecon functions.
There are 1000 observations each with 3 replicates for 3 variables V1, V2, V3. The first column represents the observations.
}
\keyword{datasets}
