% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_output.R
\name{bdeconvnc_result}
\alias{bdeconvnc_result}
\title{Constructor for \code{nc_res} Objects}
\usage{
bdeconvnc_result(obj, res_type = c("final", "mean", "all"))
}
\arguments{
\item{obj}{A list as returned by bdeconv.uninc().}

\item{res_type}{The type of result returned from the Markov chain Monte Carlo
sampler.}
}
\value{
An \code{nc_res} object.
}
\description{
Constructor for \code{nc_res} Objects
}
\seealso{
\code{\link[=bdeconv]{bdeconv()}}
}
\keyword{internal}
