% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_output.R
\name{bdeconvzi_result}
\alias{bdeconvzi_result}
\title{Constructor for \code{zi_res} Objects}
\usage{
bdeconvzi_result(obj, res_type = c("final", "mean", "all"))
}
\arguments{
\item{obj}{A list as returned by bdeconv.unizi().}

\item{res_type}{The type of result returned from the Markov chain Monte Carlo
sampler.}
}
\value{
A \code{zi_res} object.
}
\description{
Constructor for \code{zi_res} Objects
}
\seealso{
\code{\link[=bdeconv]{bdeconv()}}
}
\keyword{internal}
