% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{B-Spline}
\alias{B-Spline}
\alias{dbspline}
\alias{bspline}
\alias{pbspline}
\title{The Quadratic B-Spline Distribution}
\usage{
dbspline(x, theta, lwr, upr, log = FALSE)

pbspline(x, theta, lwr, upr, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{theta}{A vector of unnormalized spline parameters.}

\item{lwr}{A scalar lower bound.}

\item{upr}{A scalar upper bound.}

\item{log}{A logical value indication whether a probability \code{p} should be
given as \code{log(p)}}
}
\value{
\code{dbspline} gives the density. The length of the result is determined
by the length of \code{x} for \code{dbspline}.
}
\description{
Density function for the normalized quadratic B-Spline distribution.
}
\keyword{internal}
