% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline_basis}
\alias{bspline_basis}
\title{Generate a B-spline Basis Matrix}
\usage{
bspline_basis(x, lwr, upr, K)
}
\arguments{
\item{x}{A numeric vector of values on which to evaluate the basis.}

\item{lwr}{A lower bound.}

\item{upr}{An upper bound.}

\item{K}{A number of knot points.}
}
\value{
A B-spline basis matrix.
}
\description{
Generate a B-spline Basis Matrix
}
\seealso{
\link{bspline}
}
\keyword{internal}
