% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline_normcoef}
\alias{bspline_normcoef}
\title{Normalize B-Spline Parameters}
\usage{
bspline_normcoef(theta, lwr, upr)
}
\arguments{
\item{theta}{A B-spline parameter vector.}

\item{lwr}{A lower bound.}

\item{upr}{An upper bound.}
}
\value{
Normalize B-spline parameters.
}
\description{
Normalize B-Spline Parameters
}
\seealso{
\link{bspline}
}
\keyword{internal}
