% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens.R
\name{dens_u}
\alias{dens_u}
\title{Error Density Evaluation}
\usage{
dens_u(obj, vals)
}
\arguments{
\item{obj}{A \code{bdeconv_result} object.}

\item{vals}{A numeric vector (or matrix) on which to evaluate the error
density.}
}
\value{
A numeric matrix or array of density values.
}
\description{
Error Density Evaluation
}
\keyword{internal}
