% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens.R
\name{dens_x}
\alias{dens_x}
\title{Estimated density function for the true variable(s)}
\usage{
dens_x(obj, vals, expand = FALSE)
}
\arguments{
\item{obj}{A \code{bdeconv_result} object.}

\item{vals}{A numeric vector (for univariate) or matrix (for multivariate) on which to evaluate the density.}

\item{expand}{Used only for \code{mv_res} objects. A logical value indicating
whether or not to expand the matrix of values.}
}
\value{
A numeric matrix or array of density values for which each row represents the estimated density corresponding to each MCMC iteration.
}
\description{
Estimated density function for the true variable(s)
}
