% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_u}
\alias{dist_u}
\title{Error Distribution Evaluation}
\usage{
dist_u(obj, vals)
}
\arguments{
\item{obj}{A \code{bdeconv_result} object.}

\item{vals}{A numeric vector (or matrix) on which to evaluate the error
density.}
}
\value{
A numeric matrix or array of cumulative distribution values.
}
\description{
Error Distribution Evaluation
}
\keyword{internal}
