% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_x}
\alias{dist_x}
\title{Estimated cumulative distribution function for the true variable(s)}
\usage{
dist_x(obj, vals, expand = FALSE)
}
\arguments{
\item{obj}{A \code{bdeconv_result} object.}

\item{vals}{A numeric vector (for univariate) or matrix (for multivariate) on which to evaluate the CDF.}

\item{expand}{Used only for \code{mv_res} objects. A logical value indicating
whether or not to expand the matrix of values.}
}
\value{
A numeric matrix or array of cumulative distribution values for which each row represents the estimated CDF corresponding to each MCMC iteration.
}
\description{
Estimated cumulative distribution function for the true variable(s)
}
