% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_optimizers.R
\name{fr.dens}
\alias{fr.dens}
\title{Maximization function for B-spline distribution parameter initialization in
\code{\link[=bdeconv]{bdeconv()}}}
\usage{
fr.dens(thetas, xs, knots.t, P.t, s2t, density.est, K.t)
}
\arguments{
\item{thetas}{Spline parameters.}

\item{xs}{Data values.}

\item{knots.t}{Knot points.}

\item{P.t}{Penalty matrix.}

\item{s2t}{Variance}

\item{density.est}{Density estimate on grid.}

\item{K.t}{Spline cardinality.}
}
\value{
A score to be maximized.
}
\description{
Maximization function for B-spline distribution parameter initialization in
\code{\link[=bdeconv]{bdeconv()}}
}
\keyword{internal}
