% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_optimizers.R
\name{fr.prob.nonzero}
\alias{fr.prob.nonzero}
\title{Maximization function for B-spline zero-inflation probability function
parameter initialization in \code{\link[=bdeconv]{bdeconv()}}}
\usage{
fr.prob.nonzero(log_diff_thetas, xs, knots.t, P.t, s2t, probs.emp)
}
\arguments{
\item{log_diff_thetas}{Spline parameters.}

\item{xs}{Data values.}

\item{knots.t}{Knot points.}

\item{P.t}{Penalty matrix.}

\item{s2t}{Variance}

\item{probs.emp}{Empirical consumption probability.}
}
\value{
A score to be maximized.
}
\description{
Maximization function for B-spline zero-inflation probability function
parameter initialization in \code{\link[=bdeconv]{bdeconv()}}
}
\keyword{internal}
