% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_optimizers.R
\name{fr}
\alias{fr}
\title{Maximization function for variance function B-spline parameter initialization
in \code{\link[=bdeconv]{bdeconv()}}}
\usage{
fr(thetas, xs, mis, knots.t, P.t, s2t, us)
}
\arguments{
\item{thetas}{Spline parameters.}

\item{xs}{Data values.}

\item{mis}{Subject indicators.}

\item{knots.t}{Knot points.}

\item{P.t}{Penalty matrix.}

\item{s2t}{Variance.}

\item{us}{Residuals.}
}
\value{
A score to be maximized.
}
\description{
Maximization function for variance function B-spline parameter initialization
in \code{\link[=bdeconv]{bdeconv()}}
}
\keyword{internal}
