% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean.bdeconv_result}
\alias{mean.bdeconv_result}
\title{Computing mean of MCMC samples}
\usage{
\method{mean}{bdeconv_result}(x, ...)
}
\arguments{
\item{x}{A \code{bdeconv_result} object.}

\item{...}{Unused.}
}
\value{
A \code{bdeconv_result} object with one sample constituting the mean
}
\description{
Computing mean of MCMC samples
}
\details{
Where a direct point-wise mean is not appropriate (i.e. for mixture
density results), mixtures are expanded to included every component from
every sampled density with weights divided by \code{n}, the number of samples
from which the mean is being taken.
}
