% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{P1.mat}
\alias{P1.mat}
\title{Generate a Penalty Matrix}
\usage{
P1.mat(K)
}
\arguments{
\item{K}{An integer dimensino value.}
}
\value{
A first-order difference penalty matrix.
}
\description{
Generate a Penalty Matrix
}
\keyword{internal}
