% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bdeconv_result}
\alias{plot.bdeconv_result}
\title{Visualization of \code{bdeconv_result} object}
\usage{
\method{plot}{bdeconv_result}(x, ..., use_ggplot = TRUE)
}
\arguments{
\item{x}{A \code{bdeconv_result} object.}

\item{...}{Unused.}

\item{use_ggplot}{A logical value indicating whether the plots will be generated using \code{ggplot2} or using base plot functions.}
}
\value{
A collection of plots illustrating various densities and related functions.
}
\description{
Visualization of \code{bdeconv_result} object
}
\details{
By default it uses \code{ggplot2} package if it is available in the system library, otherwise it falls back to the base R plot function.
}
