% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bdeconv_result}
\alias{print.bdeconv_result}
\title{Printing summary of MCMC samples for \code{bdeconv_result} object}
\usage{
\method{print}{bdeconv_result}(x, ...)
}
\arguments{
\item{x}{A \code{bdeconv_result} object.}

\item{...}{Unused.}
}
\value{
Prints the five number summary for the relevant variables.
}
\description{
Printing summary of MCMC samples for \code{bdeconv_result} object
}
\details{
For episodic components, this gives five number summary for the observed surrogates (y), the latent surrogates corresponding to the consumption days (w), the true intake patterns (x) and the errors (u). For regular components, this gives five number summary for y, x and u only.
}
