% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_sigma.R
\name{prop.sig.thetas.fn}
\alias{prop.sig.thetas.fn}
\title{Generate a Proposal Variance Matrix for B-Spline Parameters}
\usage{
prop.sig.thetas.fn(thetas, xs, mis, us, s2t, K.t, P.t, knots.t, n)
}
\arguments{
\item{thetas}{Spine parameters.}

\item{xs}{Data values.}

\item{mis}{Subject indicators.}

\item{us}{Residuals.}

\item{s2t}{Variance.}

\item{K.t}{Spline cardinality.}

\item{P.t}{Penalty matrix.}

\item{knots.t}{Knot points.}

\item{n}{Number of grid points.}
}
\value{
A proposal matrix.
}
\description{
Generate a Proposal Variance Matrix for B-Spline Parameters
}
\keyword{internal}
