% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_result}
\alias{subset_result}
\alias{[.zi_res}
\alias{[.nc_res}
\alias{[.mv_res}
\title{Subset Methods for \code{bdeconv_result} object}
\usage{
\method{[}{zi_res}(obj, i)

\method{[}{nc_res}(obj, i)

\method{[}{mv_res}(obj, i, j)
}
\arguments{
\item{obj}{A \code{bdeconv_result} object.}

\item{i}{An integer row index (or indices) to select some specific MCMC iterations.}

\item{j}{And integer column index (or indices) to select some specific set of variables (For multivariate case only).}
}
\value{
A \code{bdeconv_result} object containing the selected iterations of posterior samples for the specified variables (in the multivariate case), suitable for further analysis and visualization.
}
\description{
Subset Methods for \code{bdeconv_result} object
}
