% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtcnorm.R
\name{var_e_fn}
\alias{var_e_fn}
\title{Variance Scaling For Mixtures of Restricted Two-Component Normals}
\usage{
var_e_fn(pi, p, mu, sig1, sig2)
}
\arguments{
\item{pi}{A vector of mixing probabilities.}

\item{p}{A vector of mixture weights as in \link{drtcnorm}.}

\item{mu}{A vector of separation parameters.}

\item{sig1}{A vector of standard deviations for the first component.}

\item{sig2}{A vector of standard deviations for the second component.}
}
\value{
A variance scaling term.
}
\description{
Variance Scaling For Mixtures of Restricted Two-Component Normals
}
\seealso{
\link{rtcnorm}
}
\keyword{internal}
