% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfun.R
\name{varfun}
\alias{varfun}
\title{Error Variance as a function of X}
\usage{
varfun(obj, vals)
}
\arguments{
\item{obj}{A \code{bdeconv_result} object.}

\item{vals}{A numeric vector (or matrix) on which to evaluate the error
variance.}
}
\value{
A numeric matrix or array of variance values.
}
\description{
Error Variance as a function of X
}
\keyword{internal}
